refactor(Systems): breakdown large cluttered Systems component
