+++ src/SmartComponents/Systems/SystemPage.test.js
@@ -2,12 +2,12 @@ import { Provider, useSelector } from 'react-redux';
 import { act } from 'react-dom/test-utils';
 import { BrowserRouter as Router } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
-import { exportSystemsCSV, exportSystemsJSON, fetchIDs } from '../../Utilities/api';
-import { systemRows } from '../../Utilities/RawDataForTesting';
+import { exportSystemsCSV, exportSystemsJSON, fetchIDs } from '../../Utilities/api.js';
+import { systemRows } from '../../Utilities/RawDataForTesting.js';
 import { initMocks, mountWithIntl } from '../../Utilities/unitTestingUtilities.js';
-import Systems from './Systems';
-import UnassignSystemsModal from '../Modals/UnassignSystemsModal';
-import NoRegisteredSystems from '../../PresentationalComponents/Snippets/NoRegisteredSystems';
+import Systems from './SystemsPage.js';
+import UnassignSystemsModal from '../Modals/UnassignSystemsModal.js';
+import NoRegisteredSystems from '../../PresentationalComponents/Snippets/NoRegisteredSystems.js';
 import { render } from '@testing-library/react';
 import { IntlProvider } from '@redhat-cloud-services/frontend-components-translations';
 import messages from '../../../locales/en.json';
